+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - 0.3.0-beta
+    - Added Vantage Point Tree for spatial queries
     - Added Bzip2 serializers
     - Added Levenshtein distance kernel
     - Move K Best Selector to main repository